requirements.txt
discord.py
discord-flags
matplotlib
sklearn
quantulum3